+++ install_packages.sh
@@ -3,8 +3,8 @@
 source utils.sh
 
 PACKAGES_FILE=$(pwd)/packages/list.lst
-print_info "Installing Software"
-apt update &>/dev/null
+print_info "Installing Software" $blueColour
+sudo apt update &>/dev/null
 if [ $? -eq 0 ];
 then
     packages=$(/usr/bin/cat $PACKAGES_FILE)
@@ -12,11 +12,21 @@ then
         package_type=$(echo $line | cut -d '|' -f1)
         package_name=$(echo $line | cut -d '|' -f2)
         if [ $package_type == 'apt' ]; then
-            print_info "Installing $package_name" 
-            sudo apt install -y $package_name
+            if ! dpkg -l |grep -q -w $package_name ;
+            then
+                print_info "[+] Installing $package_name" 
+                sudo apt install -y $package_name
+            else
+                print_info "[!] skipping $package_name" $yellowColour 
+            fi
         elif [ $package_type == 'pip' ]; then
-            print_info "Install python package $package_name"  
-            pip install $package_name
+            if ! pip show $package_name >/dev/null 2>&1;
+            then 
+                print_info "Install python package $package_name"  
+                pip install $package_name
+            else
+                print_info "[!] skipping $package_name" $magentaColour 
+            fi
         elif [ $package_type == 'git' ]; then
                 
                 destination=$(echo $line | cut -d '|' -f3)
